# Refactor Distance for Increased Performance