--- a/TheSadRogue.Primitives.UnitTests.NonThreadSafe/TheSadRogue.Primitives.UnitTests.NonThreadSafe.csproj
+++ b/TheSadRogue.Primitives.UnitTests.NonThreadSafe/TheSadRogue.Primitives.UnitTests.NonThreadSafe.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <LangVersion>8.0</LangVersion>
     <Nullable>enable</Nullable>
     <IsPackable>false</IsPackable>