--- a/TheSadRogue.Primitives.UnitTests/Serialization/ImplicitConversionTests.cs
+++ b/TheSadRogue.Primitives.UnitTests/Serialization/ImplicitConversionTests.cs
@@ -18,6 +18,10 @@ namespace SadRogue.Primitives.UnitTests.Serialization
 
         private static MethodInfo? FindImplicitConversionOnType(Type typeToScan, Type from, Type to)
         {
+            typeToScan = TestData.ImplicitTypeConversionMap.GetValueOrDefault(typeToScan, typeToScan)!;
+            from = TestData.ImplicitTypeConversionMap.GetValueOrDefault(from, from)!;
+            to = TestData.ImplicitTypeConversionMap.GetValueOrDefault(to, to)!;
+
             var methods = typeToScan.GetMethods();
             foreach (var method in methods)
             {