… TheSadRogue.Primitives.UnitTests/Serialization/Comparisons.cs