For some reason, implementing `Distance.Calculate` as a structure with a switch expression or even a switch statement is not particularly efficient in C#.  Notable performance gains can be achieved by instead making the `Distance` class abstract, then implementing `Calculate` for each distance type in a subclass.  This PR implements that concept.

# Changes
- `Distance` is now a class instead of a struct, and its `Calculate(double, double)` function is now abstract
    - Each type of distance calculation is implemented as a subclass of `Distance` which implements `Calculate` appropriately
    - NOTE: This change does _not_ enable a user to create their own distance calculations, since each calculation must still have a corresponding enum type.  This limitation remains unchanged from the previous version)
- Absolute-value is no longer performed on `dx` and `dy` for euclidean distance; it is unnecessary since the value is multiplied by itself during the calculation (which, by definition, returns a positive value)
- Implemented performance tests to demonstrate the performance increase of this implementation as opposed to alternatives
- Modified serialization unit tests to be able to map subclasses to their superclass for the purposes of finding implicit conversions to serialized types
- Updated unit tests project to .NET 6
- Added appropriate .NET workflow to run unit tests as a GitHub Action (must pass before merge); closes #77 

# Potential for Breaking Changes
The API, functionally, has not changed, although the type definition has.  One area where breaking changes are possible as a result of this change is serialization.  The different distance calculations are now separate types, as opposed to simply an instance of the `Distance` struct; so in cases where type as listed in a class is of type `Distance`, the format in which JSON would serialize it will change to include the type name.

Note that this is only a problem if someone isn't using a type converter to convert Distance to/from the corresponding `Distance.Types` value (which is, in my opinion), the preferable way to serialize a `Distance` to begin with).

# Performance Changes
The following results outline the performance of this implementation, compared to the previous one as well as some alternatives:

|                   Method | DeltaX | DeltaY | DistanceType |      Mean |     Error |    StdDev |
|------------------------- |------- |------- |------------- |----------:|----------:|----------:|
| DistanceSwitchExpression |      4 |      7 |    Manhattan | 1.6592 ns | 0.0215 ns | 0.0202 ns |
|  DistanceSwitchStatement |      4 |      7 |    Manhattan | 1.2783 ns | 0.0243 ns | 0.0216 ns |
|             DistanceFunc |      4 |      7 |    Manhattan | 0.9468 ns | 0.0328 ns | 0.0291 ns |
|                 Distance |      4 |      7 |    Manhattan | 0.4935 ns | 0.0111 ns | 0.0098 ns |
| DistanceSwitchExpression |      4 |      7 |    Euclidean | 1.6815 ns | 0.0230 ns | 0.0215 ns |
|  DistanceSwitchStatement |      4 |      7 |    Euclidean | 1.4935 ns | 0.0238 ns | 0.0199 ns |
|             DistanceFunc |      4 |      7 |    Euclidean | 1.4757 ns | 0.0037 ns | 0.0033 ns |
|                 Distance |      4 |      7 |    Euclidean | 0.5969 ns | 0.0047 ns | 0.0036 ns |
| DistanceSwitchExpression |      4 |      7 |    Chebyshev | 2.1610 ns | 0.0517 ns | 0.0484 ns |
|  DistanceSwitchStatement |      4 |      7 |    Chebyshev | 1.4665 ns | 0.0555 ns | 0.0519 ns |
|             DistanceFunc |      4 |      7 |    Chebyshev | 1.4168 ns | 0.0112 ns | 0.0100 ns |
|                 Distance |      4 |      7 |    Chebyshev | 0.9107 ns | 0.0069 ns | 0.0054 ns |

The source code for these performance tests can be found in the performance tests project, in the `DistanceDispatch.cs` file.

Clearly, the implementation labeled ``Distance`, which represents the one in the primitives library (eg. the one based on an abstract class) is faster in these tests cases.

To ensure that performance gains would translate into real-world scenarios, I compiled this version of the primitives library and installed it into GoRogue.  In `SenseMap` tests, where `Distance` calculations take up notable time, this implementation was about 5-7% faster than the original.  Additionally, `AStar` showed some minor improvement (see below):

Old Distance:
|                  Method | PathDistance | DistanceCalc |      Mean |     Error |    StdDev |
|------------------------ |------------- |------------- |----------:|----------:|----------:|
|     AStarGoRogueOpenMap |           10 |    Manhattan |  3.468 us | 0.0488 us | 0.0457 us |
| FastAStarGoRogueOpenMap |           10 |    Manhattan |  3.500 us | 0.0456 us | 0.0426 us |
|     AStarGoRogueOpenMap |           10 |    Euclidean |  5.443 us | 0.0927 us | 0.0867 us |
| FastAStarGoRogueOpenMap |           10 |    Euclidean |  5.481 us | 0.0652 us | 0.0545 us |
|     AStarGoRogueOpenMap |           10 |    Chebyshev |  5.411 us | 0.0227 us | 0.0201 us |
| FastAStarGoRogueOpenMap |           10 |    Chebyshev |  5.350 us | 0.0421 us | 0.0394 us |
|     AStarGoRogueOpenMap |           50 |    Manhattan | 19.757 us | 0.0980 us | 0.0818 us |
| FastAStarGoRogueOpenMap |           50 |    Manhattan | 19.745 us | 0.1632 us | 0.1447 us |
|     AStarGoRogueOpenMap |           50 |    Euclidean | 29.870 us | 0.4062 us | 0.3800 us |
| FastAStarGoRogueOpenMap |           50 |    Euclidean | 29.917 us | 0.5460 us | 0.5108 us |
|     AStarGoRogueOpenMap |           50 |    Chebyshev | 28.331 us | 0.1362 us | 0.1207 us |
| FastAStarGoRogueOpenMap |           50 |    Chebyshev | 28.508 us | 0.0970 us | 0.0908 us |
|     AStarGoRogueOpenMap |          100 |    Manhattan | 42.732 us | 0.1990 us | 0.1862 us |
| FastAStarGoRogueOpenMap |          100 |    Manhattan | 43.417 us | 0.1830 us | 0.1529 us |
|     AStarGoRogueOpenMap |          100 |    Euclidean | 60.831 us | 0.1879 us | 0.1666 us |
| FastAStarGoRogueOpenMap |          100 |    Euclidean | 60.385 us | 0.2839 us | 0.2516 us |
|     AStarGoRogueOpenMap |          100 |    Chebyshev | 59.157 us | 0.2601 us | 0.2172 us |
| FastAStarGoRogueOpenMap |          100 |    Chebyshev | 58.368 us | 0.2545 us | 0.2125 us |

New Distance:
|                  Method | PathDistance | DistanceCalc |      Mean |     Error |    StdDev |
|------------------------ |------------- |------------- |----------:|----------:|----------:|
|     AStarGoRogueOpenMap |           10 |    Manhattan |  3.199 us | 0.0108 us | 0.0090 us |
| FastAStarGoRogueOpenMap |           10 |    Manhattan |  3.239 us | 0.0148 us | 0.0131 us |
|     AStarGoRogueOpenMap |           10 |    Euclidean |  5.032 us | 0.0485 us | 0.0405 us |
| FastAStarGoRogueOpenMap |           10 |    Euclidean |  5.020 us | 0.0922 us | 0.0817 us |
|     AStarGoRogueOpenMap |           10 |    Chebyshev |  5.085 us | 0.0540 us | 0.0479 us |
| FastAStarGoRogueOpenMap |           10 |    Chebyshev |  5.000 us | 0.0882 us | 0.0825 us |
|     AStarGoRogueOpenMap |           50 |    Manhattan | 18.829 us | 0.2184 us | 0.1824 us |
| FastAStarGoRogueOpenMap |           50 |    Manhattan | 18.918 us | 0.3107 us | 0.2755 us |
|     AStarGoRogueOpenMap |           50 |    Euclidean | 27.424 us | 0.1965 us | 0.1742 us |
| FastAStarGoRogueOpenMap |           50 |    Euclidean | 27.686 us | 0.1658 us | 0.1551 us |
|     AStarGoRogueOpenMap |           50 |    Chebyshev | 27.184 us | 0.4351 us | 0.4070 us |
| FastAStarGoRogueOpenMap |           50 |    Chebyshev | 26.934 us | 0.3349 us | 0.2969 us |
|     AStarGoRogueOpenMap |          100 |    Manhattan | 41.617 us | 0.7003 us | 0.6550 us |
| FastAStarGoRogueOpenMap |          100 |    Manhattan | 41.493 us | 0.6014 us | 0.5332 us |
|     AStarGoRogueOpenMap |          100 |    Euclidean | 57.502 us | 0.9129 us | 0.8539 us |
| FastAStarGoRogueOpenMap |          100 |    Euclidean | 57.471 us | 0.2636 us | 0.2337 us |
|     AStarGoRogueOpenMap |          100 |    Chebyshev | 55.681 us | 0.3584 us | 0.2992 us |
| FastAStarGoRogueOpenMap |          100 |    Chebyshev | 56.146 us | 0.8530 us | 0.7979 us |